add dual mode

+++ udp/define.h
@@ -18,8 +18,8 @@ struct board{
     int status;
     char* time;
     int map_info;
-    int pos_x;
-    int pos_y;
+    float pos_x;
+    float pos_y;
     float angle;
     struct sockaddr_in board_addr;
 };
